[Does not work yet] UdfpsHelper: Track brightness mirror to hide/show Udfps dim layer
This provides a much deeper integration with Udfps dim layer such
that we don't need to look for APIs that can track temporary brightness.
That is because when the brightness mirror is shown, there is no necessity
for the Udfps dim layer to be present.

It also fixes the issue where dim layer is hidden with QS expansion -> Device Control
where the alternate bouncer is brought up.

Change-Id: I53f78cbcfd3b42a3b5dbe6110424687ebfac102d